a6 completed

+++ Kambaz/Enrollments/dao.js
@@ -1,37 +1,19 @@
-// import Database from "../Database/index.js";
-// import { v4 as uuidv4 } from "uuid";
-
-// export function enrollUserInCourse(userId, courseId) {
-//     const { enrollments } = Database;
-//     enrollments.push({ _id: uuidv4(), user: userId, course: courseId });
-// }
-
-// export function deleteEnrollment(userId, courseId) {
-//     const { enrollments } = Database;
-//     Database.enrollments = enrollments.filter((enrollment) => enrollment.user !== userId || enrollment.course !== courseId);
-// }
-
-// export function getAllEnrollments() {
-//     const { enrollments } = Database;
-//     return enrollments;
-// }
-
 import model from "./model.js";
+import { v4 as uuidv4 } from "uuid";
 
-export async function findCoursesForUser(userId) {
-    const enrollments = await model.find({ user: userId }).populate("course");
-    console.log("enroollments");
-    console.log(userId);
-    console.log(enrollments);
-    return enrollments.map((enrollment) => enrollment.course);
+export function enrollUserInCourse(userId, courseId) {
+    const newEnrollment = { course: courseId, user: userId, _id: uuidv4() };
+    return model.create(newEnrollment);
 }
-export async function findUsersForCourse(courseId) {
-    const enrollments = await model.find({ course: courseId }).populate("user");
-    return enrollments.map((enrollment) => enrollment.user);
+
+export function deleteEnrollment(userId, courseId) {
+    return model.deleteOne({ user: userId, course: courseId });
 }
-export function enrollUserInCourse(user, course) {
-    return model.create({ user, course, _id: `${user}-${course}` });
+
+export function getAllEnrollments() {
+    return model.find();
 }
-export function unenrollUserFromCourse(user, course) {
-    return model.deleteOne({ user, course });
+
+export function findUsersForCourse(courseId) {
+    return model.find({ course: courseId }).populate("user");
 }
